refactored so game options are contained within their own packages
refactored so game options are contained within their own packages
