+++ src/test/scala/behaviors/controller/BlackjackControllerSpec.scala
@@ -11,8 +11,8 @@ import cards.classes.Rank._
 import cards.classes.Suit._
 import cards.classes.hand.Hand
 import cards.classes.state.{ BlackjackGameState, BlackjackPlayerState }
-import cards.classes.options.BlackjackOptions
-import cards.classes.options.BlackjackPayout._
+import cards.classes.options.blackjack.BlackjackOptions
+import cards.classes.options.blackjack.BlackjackPayout._
 import cards.classes.actions.{ Action, BlackjackAction }
 import cards.classes.actions.BlackjackAction._
 import org.scalatest.flatspec.AnyFlatSpec
