+++ src/test/scala/behaviors/play/BlackjackPlaySpec.scala
@@ -8,10 +8,10 @@ import cards.classes.DeckType._
 import cards.classes.Rank._
 import cards.classes.Suit._
 import cards.classes.hand.Hand
-import cards.classes.options.BlackjackOptions
-import cards.classes.options.DealerHitLimit._
+import cards.classes.options.blackjack.BlackjackOptions
+import cards.classes.options.blackjack.DealerHitLimit._
+import cards.classes.options.blackjack.BlackjackPayout._
 import cards.classes.state.{ BlackjackGameState, BlackjackPlayerState }
-import cards.classes.options.BlackjackPayout._
 import cards.classes.actions.{ Action, BlackjackAction }
 import cards.classes.actions.BlackjackAction._
 import org.scalatest.flatspec.AnyFlatSpec
